# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LunamHub Family Command Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /redemptions:
    get:
      operationId: listRedemptions
      tags: [rewards]
      summary: List reward redemption requests
      responses:
        "200":
          description: Redemptions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redemption"
    post:
      operationId: requestRedemption
      tags: [rewards]
      summary: Request to redeem a reward (child action)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedemptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/approve:
    post:
      operationId: approveRedemption
      tags: [rewards]
      summary: Approve a redemption request (parent action, deducts points)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/reject:
    post:
      operationId: rejectRedemption
      tags: [rewards]
      summary: Reject a redemption request (parent action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"

  # ── Lists ─────────────────────────────────────────────────────────────────────
  /lists:
    get:
      operationId: listLists
      tags: [lists]
      summary: List all shared lists
      responses:
        "200":
          description: Lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharedList"
    post:
      operationId: createList
      tags: [lists]
      summary: Create a shared list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedList"
  /lists/{id}:
    get:
      operationId: getList
      tags: [lists]
      summary: Get a list with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedListWithItems"
        "404":
          description: Not found
    patch:
      operationId: updateList
      tags: [lists]
      summary: Update a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedListUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedList"
    delete:
      operationId: deleteList
      tags: [lists]
      summary: Delete a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lists/{listId}/items:
    post:
      operationId: createListItem
      tags: [lists]
      summary: Add an item to a list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItem"
  /lists/{listId}/items/{itemId}:
    patch:
      operationId: updateListItem
      tags: [lists]
      summary: Update a list item
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItem"
    delete:
      operationId: deleteListItem
      tags: [lists]
      summary: Delete a list item
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Meals ─────────────────────────────────────────────────────────────────────
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List saved meals
      responses:
        "200":
          description: Meals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    patch:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /meal-plan:
    get:
      operationId: getMealPlan
      tags: [meals]
      summary: Get meal plan for a week
      parameters:
        - name: weekStart
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Meal plan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlanEntry"
    post:
      operationId: addMealPlanEntry
      tags: [meals]
      summary: Add a meal to the plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanEntry"
  /meal-plan/{id}:
    delete:
      operationId: deleteMealPlanEntry
      tags: [meals]
      summary: Remove a meal plan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /meal-plan/{id}/add-to-grocery:
    post:
      operationId: addMealIngredientsToGrocery
      tags: [meals]
      summary: Add meal ingredients to grocery list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingredients added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
                required:
                  - added

  # ── Routines ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FamilyMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
        role:
          type: string
          enum: [parent, child]
        pointsBalance:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - emoji
        - color
        - role
        - pointsBalance
        - createdAt
...
    Reward:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        pointsCost:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - pointsCost
        - active
        - createdAt
...
    Redemption:
      type: object
      properties:
        id:
          type: integer
        rewardId:
          type: integer
        reward:
          $ref: "#/components/schemas/Reward"
        memberId:
          type: integer
        member:
          $ref: "#/components/schemas/FamilyMember"
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - rewardId
        - memberId
        - status
        - createdAt
    RedemptionInput:
      type: object
      properties:
        rewardId:
          type: integer
        memberId:
          type: integer
      required:
        - rewardId
        - memberId

    # ── Lists ─────────────────────────────────────────────────────────────────
    SharedList:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [grocery, packing, school, reminders, other]
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - createdAt
    SharedListInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [grocery, packing, school, reminders, other]
      required:
        - name
        - category
    SharedListUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [grocery, packing, school, reminders, other]
    SharedListWithItems:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListItem"
      required:
        - id
        - name
        - category
        - createdAt
        - items
    ListItem:
      type: object
      properties:
        id:
          type: integer
        listId:
          type: integer
        text:
          type: string
        completed:
          type: boolean
        assignedTo:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - listId
        - text
        - completed
        - createdAt
    ListItemInput:
      type: object
      properties:
        text:
          type: string
        assignedTo:
          type: integer
        category:
          type: string
      required:
        - text
    ListItemUpdate:
      type: object
      properties:
        text:
          type: string
        completed:
          type: boolean
        assignedTo:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]

    # ── Meals ─────────────────────────────────────────────────────────────────
    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    MealInput:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
        ingredients:
          type: string
      required:
        - name
    MealUpdate:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
        ingredients:
          type: string
    MealPlanEntry:
      type: object
      properties:
        id:
          type: integer
        mealId:
          type: integer
        meal:
          $ref: "#/components/schemas/Meal"
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
      required:
        - id
        - mealId
        - date
        - mealType
    MealPlanEntryInput:
      type: object
      properties:
        mealId:
          type: integer
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
      required:
        - mealId
        - date
        - mealType

    # ── Routines ──────────────────────────────────────────────────────────────
...