# lib/api-zod/src/generated/types/redemption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LunamHub Family Command Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { FamilyMember } from './familyMember';
import type { RedemptionStatus } from './redemptionStatus';
import type { Reward } from './reward';

export interface Redemption {
  id: number;
  rewardId: number;
  reward?: Reward;
  memberId: number;
  member?: FamilyMember;
  status: RedemptionStatus;
  createdAt: string;
}