# lib/api-zod/src/generated/types/sharedListUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LunamHub Family Command Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { SharedListUpdateCategory } from './sharedListUpdateCategory';

export interface SharedListUpdate {
  name?: string;
  category?: SharedListUpdateCategory;
}